move and lint test
